+++ contest/language/nl/contest_acp.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * phpBB Gallery Contest ACP language.
+ *
+ * @package   phpbbgallery/contest
+ * @copyright 2018- Leinad4Mind
+ * @license   GPL-2.0-only
+ */
+
+if (!defined('IN_PHPBB'))
+{
+	exit;
+}
+
+$lang = array_merge($lang, [
+	'CONTEST_CREATION'                    => 'Nieuwe wedstrijden toestaan',
+	'CONTEST_CREATION_EXPLAIN'            => 'Staat beheerders toe nieuwe wedstrijdalbums te maken. Bestaande wedstrijden blijven actief en bewerkbaar wanneer deze optie is uitgeschakeld.',
+	'CONTEST_CREATION_DISABLED'           => 'Het maken van nieuwe wedstrijdalbums is uitgeschakeld in de Galerijconfiguratie.',
+]);