+++ contest/language/ru/contest_acp.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * phpBB Gallery Contest ACP language.
+ *
+ * @package   phpbbgallery/contest
+ * @copyright 2018- Leinad4Mind
+ * @license   GPL-2.0-only
+ */
+
+if (!defined('IN_PHPBB'))
+{
+	exit;
+}
+
+$lang = array_merge($lang, [
+	'CONTEST_CREATION'                     => 'Разрешить новые конкурсы',
+	'CONTEST_CREATION_EXPLAIN'             => 'Позволяет администраторам создавать новые конкурсные альбомы. Существующие конкурсы остаются активными и доступными для редактирования, когда эта настройка отключена.',
+	'CONTEST_CREATION_DISABLED'            => 'Создание новых конкурсных альбомов отключено в настройках Галереи.',
+]);